+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Fix error on count assertions
 - Added pipeline to add contributors to the readme
 - Added documentation with vitepress
+- Added `setUp`, `setUpBeforeScript`, `tearDown` and `tearDownAfterScript` function execution before and/or after test and/or script execution
 
 ### 0.5.0
 ### 2023-09-10